+++ additional_addons/Edumedia_India/__init__.py
@@ -15,7 +15,8 @@ import ed_hr_attendance
 import ed_leave 
 import ed_expense
 import ed_payroll
-import wizard 
+import ed_business_dev
+import wizard
